+++ react_on_rails/lib/generators/react_on_rails/templates/base/tailwind/app/javascript/src/HelloWorld/ror_components/HelloWorld.client.jsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react';
-import '../../../stylesheets/application.css';
 
 const HelloWorld = (props) => {
   const [name, setName] = useState(props.name);